+++ CHANGELOG.md
@@ -18,6 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - OpenAPI spec for `/v1` served at `/v1/openapi.json` with consolidated operations.
 - `distclean` target to remove ignored files via `git clean -fdX`.
 
+### Changed
+- Organized SmartGPT Bridge routes into versioned directories.
+
 ### Fixed
 
 - SmartGPT Bridge file actions now treat leading '/' as the repository root.
