+++ services/ts/smartgpt-bridge/src/fastifyApp.js
@@ -6,13 +6,13 @@ import swagger from '@fastify/swagger';
 import swaggerUi from '@fastify/swagger-ui';
 import ajvformats from 'ajv-formats';
 import { createFastifyAuth } from './fastifyAuth.js';
-import { registerV0Routes } from './routes/v0.js';
+import { registerV0Routes } from './routes/v0/index.js';
 
 import { indexerManager } from './indexer.js';
 import { restoreAgentsFromStore } from './agent.js';
 import { registerSinks } from './sinks.js';
 import { registerRbac } from './rbac.js';
-import { registerV1Routes } from './routes/v1.js';
+import { registerV1Routes } from './routes/v1/index.js';
 import { mongoChromaLogger } from './logging/index.js';
 
 export function buildFastifyApp(ROOT_PATH) {
